perf: move DownloadService URL-selection tests out of Robolectric
DownloadServiceTest was the suite's single slowest class (60.8s, 10.6%
of summed test time) yet ~10 of its 16 tests (getNextUrl /
getNextPriorityUrl) exercise pure companion-object logic against a
mocked SharedPreferences and touch no Android framework. They were
paying the full Robolectric + SDK-34 framework-load tax for nothing.

Extract those 10 tests into a new plain-JUnit
DownloadServiceUrlSelectionTest (no @RunWith(RobolectricTestRunner)),
leaving only the 6 startService tests under Robolectric.

Measured on the full suite (1155 tests, all still pass, 0 failures):
- DownloadServiceTest:        60.8s -> 38.0s
- DownloadServiceUrlSelectionTest (new): 0.8s
- summed test time:           574.6s -> 565.4s
